chore(api): move smart random services in evaluation context
