+++ api/src/evaluation/domain/usecases/index.js
@@ -1,7 +1,6 @@
 import { dirname, join } from 'node:path';
 import { fileURLToPath } from 'node:url';
 
-import * as smartRandomService from '../../../../lib/domain/services/algorithm-methods/smart-random.js';
 import * as improvementService from '../../../../lib/domain/services/improvement-service.js';
 import * as campaignParticipationRepository from '../../../../lib/infrastructure/repositories/campaign-participation-repository.js';
 import * as campaignRepository from '../../../../lib/infrastructure/repositories/campaign-repository.js';
@@ -24,6 +23,7 @@ import { repositories } from '../../infrastructure/repositories/index.js';
 import * as stageAcquisitionRepository from '../../infrastructure/repositories/stage-acquisition-repository.js';
 import * as stageCollectionForTargetProfileRepository from '../../infrastructure/repositories/stage-collection-repository.js';
 import * as stageRepository from '../../infrastructure/repositories/stage-repository.js';
+import * as smartRandomService from '../services/algorithm-methods/smart-random.js';
 import { getCompetenceLevel } from '../services/get-competence-level.js';
 import * as scorecardService from '../services/scorecard-service.js';
