+++ api/tests/evaluation/unit/domain/services/algorithm-methods/cat-algorithm_test.js
@@ -1,9 +1,9 @@
 import {
   findMaxRewardingSkills,
   getPredictedLevel,
-} from '../../../../../lib/domain/services/algorithm-methods/cat-algorithm.js';
-import { expect } from '../../../../test-helper.js';
-import { buildKnowledgeElement, buildSkill, buildTube } from '../../../../tooling/domain-builder/factory/index.js';
+} from '../../../../../../src/evaluation/domain/services/algorithm-methods/cat-algorithm.js';
+import { expect } from '../../../../../test-helper.js';
+import { buildKnowledgeElement, buildSkill, buildTube } from '../../../../../tooling/domain-builder/factory/index.js';
 
 describe('Unit | Domain | services | cat-algorithm', function () {
   let knowledgeElements;
