+++ api/tests/evaluation/unit/domain/services/smart-random/skills-filter_test.js
@@ -1,12 +1,12 @@
-import { Tube } from '../../../../../lib/domain/models/Tube.js';
+import { Tube } from '../../../../../../lib/domain/models/Tube.js';
+import { SmartRandomDetails } from '../../../../../../src/evaluation/domain/models/SmartRandomDetails.js';
 import {
   focusOnDefaultLevel,
   getFilteredSkillsForFirstChallenge,
   getFilteredSkillsForNextChallenge,
-} from '../../../../../lib/domain/services/algorithm-methods/skills-filter.js';
-import { SmartRandomDetails } from '../../../../../src/evaluation/domain/models/SmartRandomDetails.js';
-import { domainBuilder, expect } from '../../../../test-helper.js';
-import { buildSkill } from '../../../../tooling/domain-builder/factory/index.js';
+} from '../../../../../../src/evaluation/domain/services/algorithm-methods/skills-filter.js';
+import { domainBuilder, expect } from '../../../../../test-helper.js';
+import { buildSkill } from '../../../../../tooling/domain-builder/factory/index.js';
 
 const KNOWLEDGE_ELEMENT_STATUS = {
   VALIDATED: 'validated',
